added MokkaDBConfig init.sh script

+++ StandardConfig/scripts/rec/job.sh
@@ -316,28 +316,34 @@ test $? -eq 0 || msg CRITICAL 70 "failed to pass basic system tests"
 
 
 # ------- setup ilcsoft ------------------------------------------------------
-# download an ilcsoft version not yet installed on the grid
-tarball=ilcsoft-$SW_VER-$ARCH-full.tar.gz
-if [ ! -d ilcsoft ] ; then
-    if [ ! -e "$tarball" ] ; then
-        msg INFO "downloading ilcsoft tarball..."
-        wget "http://ilcsoft.desy.de/ilcsoft-bin-releases/$tarball"
-        test $? -eq 0 || msg CRITICAL 71 "failed to download ilcsoft"
-    fi
-    msg INFO "unpacking ilcsoft tarball..."
-    tar -xzf $tarball
-    test $? -eq 0 || msg CRITICAL 71 "failed to unpack ilcsoft tarball"
-    rm -f $tarball
-fi
-export VO_ILC_SW_DIR=$PWD
+ILCSOFT="$VO_ILC_SW_DIR/ilcsoft/$ARCH"
+# check if required ilcsoft version is available on the grid
+if [ ! -r "$ILCSOFT/$SW_VER" ] ; then
 
-# msg INFO "applying patch..."
-#tarball=ilcsoft-$SW_VER-$ARCH-patch-vormwald-0001.tgz
-#wget "http://ilcsoft.desy.de/data/production/patches/$tarball" && tar -xzvf $tarball && rm -f $tarball
-#test $? -eq 0 || msg CRITICAL 71 "failed to initialize ilcsoft"
+    # download an ilcsoft version not yet installed on the grid
+    tarball=ilcsoft-$SW_VER-$ARCH-full.tar.gz
+    if [ ! -d ilcsoft ] ; then
+        if [ ! -e "$tarball" ] ; then
+            msg INFO "downloading ilcsoft tarball..."
+            wget "http://ilcsoft.desy.de/ilcsoft-bin-releases/$tarball"
+            test $? -eq 0 || msg CRITICAL 71 "failed to download ilcsoft"
+        fi
+        msg INFO "unpacking ilcsoft tarball..."
+        tar -xzf $tarball
+        test $? -eq 0 || msg CRITICAL 71 "failed to unpack ilcsoft tarball"
+        rm -f $tarball
+    fi
+
+    #msg INFO "applying patch..."
+    #tarball=ilcsoft-$SW_VER-$ARCH-patch-0001.tgz
+    #wget "http://ilcsoft.desy.de/data/production/patches/$tarball" && tar -xzvf $tarball && rm -f $tarball
+    #test $? -eq 0 || msg CRITICAL 71 "failed to download ilcsoft patch"
+
+    export ILCSOFT=$PWD
+fi
 
 msg INFO "initialize ilcsoft..."
-. $VO_ILC_SW_DIR/ilcsoft/$ARCH/init_ilcsoft.sh $SW_VER
+. $ILCSOFT/init_ilcsoft.sh $SW_VER
 test $? -eq 0 || msg CRITICAL 71 "failed to initialize ilcsoft"
 # ----------------------------------------------------------------------------
